rename userKey to localUserSecret
