+++ app/core/UKYC/index.ts
@@ -1,8 +1,8 @@
 export {
-  getOrCreateUserKey,
-  loadUserKey,
-  hasUserKey,
-} from './userKey';
+  getOrCreateLocalUserSecret,
+  loadLocalUserSecret,
+  hasLocalUserSecret,
+} from './localUserSecret';
 export {
   deriveClientMaterial,
   encodeClientMaterial,
@@ -12,8 +12,8 @@ export type {
   EncodedUkycClientMaterial,
 } from './deriveClientMaterial';
 export {
-  UKYC_USER_KEY_PATH,
-  UKYC_USER_KEY_SIZE_BYTES,
+  UKYC_LOCAL_USER_SECRET_PATH,
+  UKYC_LOCAL_USER_SECRET_SIZE_BYTES,
   UKYC_DERIVED_KEY_SIZES,
   UKYC_KDF_INFO,
 } from './constants';
